Fixed some Stupid decisions...

+++ ComboKeypad/Keypad.ino
@@ -18,16 +18,16 @@ int count = 0;   //How many buttens that have been pressed
 int input[numOfCombos]; //Saves what the user has clicked
 //
 
-const short int del = 10000;   //Timer(10s), if user inputs a number
+const int del = 10000;   //Timer(10s), if user inputs a number
                                //and not inputs another within 10 seconds program will reset
 unsigned long time;
-short int buttonsRead[numOfButtons];   //Stores the "digitalRead()" of buttons[]
-const short int buttonStream = 2;   //Electricity for the buttons, didnt really know how to name this one so buttonStream it is
-const short int rightLight = 9;   //Correct combo PIN, "Green" light
-const short int wrongLight = 8;   //Incorrect combo PIN, "Red" Light
+int buttonsRead[numOfButtons];   //Stores the "digitalRead()" of buttons[]
+const int buttonStream = 2;   //Electricity for the buttons, didnt really know how to name this one so buttonStream it is
+const int rightLight = 9;   //Correct combo PIN, "Green" light
+const int wrongLight = 8;   //Incorrect combo PIN, "Red" Light
 ///
 /*----(Function prototyping)----*/
-void pressButton(short int num, short int digRead);
+void pressButton(int num, int digRead);
 void reset();
 void debug();
 
@@ -35,7 +35,7 @@ void debug();
 /*----(Function setup)----*/
 void setup() {
     //Used to be able to add as many buttons as you want
-    for(short int setb = 0; setb < numOfButtons; setb++){
+    for(int setb = 0; setb < numOfButtons; setb++){
       pinMode(buttons[setb], INPUT); // sets All Input pin
      }
 
@@ -53,7 +53,7 @@ void setup() {
 /*----(Function loop)----*/
 void loop() {
     //Gets the Button States
-for(short int state = 0; state < numOfButtons; state++){
+for(int state = 0; state < numOfButtons; state++){
   buttonsRead[state] = digitalRead(buttons[state]); // First Switch
 }
 
@@ -66,8 +66,8 @@ for(short int state = 0; state < numOfButtons; state++){
       reset();
     }
 
-short int value = 0;   //how many combos are correct
-for(short int a = 0; a < numOfCombos; a++){
+int value = 0;   //how many combos are correct
+for(int a = 0; a < numOfCombos; a++){
     if(input[a] == combination[a]){   //Checks if the input and the combo are the same
      value++;   //Adds one if combo is correct
     }
@@ -90,7 +90,7 @@ for(short int a = 0; a < numOfCombos; a++){
 /// this for loop is here to be able to customize this code                 ///
 /// to add any amount of combinations and buttons                           ///
 ///////////////////////////////////////////////////////////////////////////////
-for(short int a = 0; a < numOfCombos; a++){
+for(int a = 0; a < numOfCombos; a++){
   if(input[a] != combination[a] && count == numOfCombos){
     digitalWrite(wrongLight, 1);   //Lights up the "red" button
      Serial.println("Wrong");   //Prints that the combination is wrong
@@ -114,16 +114,16 @@ if (count < numOfCombos) { // Restricts the variable to not go over 4
 //////////////////////////////////////////////////////
 
 //Function pressButton, If user presses a Button
-void pressButton(short int num, short int digRead){
+void pressButton(int num, int digRead){
   if (digRead == 1 && doonce != true) { // If I hit a Button
    time = millis();   //Starts the 10 second timer
     input[count] = num + 1; // sends the number 1 into the array at the "count number"
      count++; // Where to put 4 into the array
       doonce = true;
     }
-    short int exists;
+    int exists;
 
-      for(short int a = 0; a < numOfButtons; a++){
+      for(int a = 0; a < numOfButtons; a++){
         if(buttonsRead[a] == 0){
           exists++;
         }
@@ -141,7 +141,7 @@ void pressButton(short int num, short int digRead){
 
 /*----(Function Reset)----*/
 void reset() {
-  for (short int inputIndex = 0; inputIndex < numOfCombos; inputIndex++) {
+  for (int inputIndex = 0; inputIndex < numOfCombos; inputIndex++) {
     input[inputIndex] = 0;   //Zeros the input array
   }
 
@@ -158,7 +158,7 @@ void debug(){
    Serial.print(count);
     Serial.print(", |");
 
-     for(short int a = 0; a < numOfCombos; a++){
+     for(int a = 0; a < numOfCombos; a++){
       Serial.print(input[a]);   //Prints which number has been pressed
     }
   Serial.println();
